LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/border-image-repeat-expected.txt


PASS Can set 'border-image-repeat' to CSS-wide keywords
FAIL Can set 'border-image-repeat' to var() references assert_equals: expected 2 but got 1
FAIL Can set 'border-image-repeat' to the 'stretch' keyword assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'border-image-repeat' to the 'repeat' keyword assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'border-image-repeat' to the 'round' keyword assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'border-image-repeat' to the 'space' keyword assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
PASS Setting 'border-image-repeat' to a length throws TypeError
PASS Setting 'border-image-repeat' to a percent throws TypeError
PASS Setting 'border-image-repeat' to a time throws TypeError
PASS Setting 'border-image-repeat' to an angle throws TypeError
PASS Setting 'border-image-repeat' to a flexible length throws TypeError
PASS Setting 'border-image-repeat' to a number throws TypeError
PASS Setting 'border-image-repeat' to a URL throws TypeError
PASS Setting 'border-image-repeat' to a transform throws TypeError
PASS 'border-image-repeat' does not supported 'stretch repeat'
PASS 'border-image-repeat' does not supported 'round space'
